pyproject.toml
[tool.poetry]
name = "xl"
version = "2.0.0"
description = ""
authors = ["Artasov <ivanhvalevskey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

# Tier 1
Django = "5.1.4"
adjango = "0.3.6"
djangorestframework = "3.15.2"
adrf = "0.1.9"
celery = "5.5.0rc4"
daphne = "4.1.2"
channels = "4.2.0"

# Tier 2
django-redis = "5.4.0"
django-cors-headers = "4.6.0"
django-logui = "0.1.2"
django-endpoints = "0.0.3"
django-extensions = "3.2.3"
django-polymorphic = "3.1.0"
djangorestframework-simplejwt = "5.4.0"
django-minio-backend = "3.8.0"
django-celery-beat = "2.7.0"
django-cachalot = "2.7.0"
django-silk = "5.3.2"
django-imagekit = "5.0.0"
django-user-agents = "0.4.0"
django-object-actions = "4.3.0"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^8.0.0" }
phonenumbers = "8.13.53"
django-jazzmin = "3.0.1"
django-import-export = { version = "4.3.4", extras = ["all"] }

flower = "2.0.1"
drf-yasg = "1.21.8"
redis = "5.2.1"
channels-redis = "4.2.1"
twisted = { version = "24.11.0", extras = ["http2", "tls"] }
httpx = "1.0.0b0"
uuid6 = "2024.7.10"

aiohttp = "3.11.10"
chardet = "5.2.0"
python-dotenv = "1.0.1"
celery-singleton = "0.3.1"

Pillow = "11.0.0"
psycopg2-binary = "2.9.10"
requests = "2.32.3"
xlwt = "1.3.0"
setuptools = "75.6.0"
pyperclip = "1.9.0"
psutil= "6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
